+++ index.ts
@@ -1,6 +1,6 @@
 import { AppRegistry } from "react-native"
 
 import { name as appName } from "./app.json"
-import App from "./src/App"
+import App from "./src/OldApp"
 
 AppRegistry.registerComponent(appName, () => App)
