+++ src/OldApp/navigators/RestaurantsNavigator/RestaurantsNavigator.tsx
