+++ src/OldApp/navigators/RestaurantsNavigator/index.ts
