+++ src/screens/RestaurantDetails/RestaurantDetails.tsx
@@ -2,7 +2,7 @@ import { useNavigation } from "@react-navigation/native"
 import { StackScreenProps } from "@react-navigation/stack"
 import { useEffect } from "react"
 
-import { RestaurantsStackParamList } from "../../App"
+import { RestaurantsStackParamList } from "../../OldApp"
 import Loading from "../../shared/components/Loading"
 import RestaurantHeader from "../../shared/components/RestaurantHeader"
 import Button from "../../shared/design/Button"
